Give escaped newlines meaning
Previously, escaped newlines would be inserted into text as raw newlines.
It doesn't make sense to insert a newline into sentences,
so this commit changes the behaviour to treat an escaped newline
as "please continue the sentence" as opposed to normal newlines
which break sentences.

This also merits the inclusion of a new error, because escaped newlines
only have meaning in paragraph mode.
Finding one in block mode doesn't make sense.