Remove hashes from inline/block scopes and code
I want inline and block scope opens to merge into codeclose,
so the user has to express intent for what the codeclose should be.
This means getting rid of the hashes in codeopen, codeclose, scopeopens, and scopeclose.
RawScopeOpen and RawScopeClose (a new token) are now the only tokens that
use hashes - they need to do this for raw text.
CodeOpen and CodeClose use multiple square brackets [[[ ]]] instead of hashes.
Normal scopes (inline or block) don't use hashes at all.