Manual Inline checking
The use case for "Inline" being literally any non-block thing is iffy.
e.g. under that definition, Sentence and List[Inline] are Inline and can be recursive.
Generally using [1+2] in your code to insert "3" isn't needed, which was a primary usecase.
It's better to have the typesafety of "I know the thing I'm interacting with
is supposed to be inline."

This commit makes the Inline typeclass opt-in, like Block.
It also adds InlineScope, like BlockScope.
build_from_inlines for InlineScopeBuilder now takes InlineScope and returns Inline.

When Inline objects are expected, they are immediately casted
so the error can appear when the code span is known.