Move to "dictified object" renderer model
Because we don't create a separate interpreter for each file now,
we need a way to encapsulate the state of a renderer per-file.
The correct way to do that in python is via objects
(i.e. create a Renderer object with state like footnotes/citations)
but this has ergonomics issues: we don't want
to have to specify "renderer.footnote" in turniptext code.
Solution: "dictify" the renderer object into a dict of bound methods
and fields,
and use it as the "locals" dict when execing python from rust.
Special care is needed when using properties,
because they can't be stored directly by dicts and won't be invoked
multiple times => if they're impure, they won't work.