Specific error for raw scopes incorrectly emitting blocks
Before, we had a function for directly asserting a value returned by a
RawScopeBuilder must be Inline, only called in the context of a paragraph.
This led to a confusing/not-automatically-testable error, which would
also conflict with the Python protocol and give the impression
the code was wrong, when really it was the usage in that place.
Add a proper error for that.