Better indent handling in renderers
Before, all sentence/paragraph breaks would immediately be followed with the indent.
This was bad because you couldn't emit a newline, change the indent,
and have further things continue with that indent.
Now the indent is only emitted on the next emit_raw after a newline,
so indent changes are more intuitive.