[WIP] Rename cli, remove parse_{file,str} Rust functions
turnip_text is no longer a Rust CLI tool, so the name is inaccurate
and these functions should be moved to be more Python-friendly.
Additionally, for parsing docsegments inside subfiles we need to keep
parser state around, so creating a new
parser inside Rust for every subfile is bad.
The old cli functions enforce this, so they must go.