[WIP] Big big changes to parser structure
Preparing for subfiles by splitting structure-related functions and data
out of InterpState and into InterpDocumentStructure
and making block stacks per-subfile.
This means you won't be able to start block scopes in a subfile without closing them,
and you won't be able to close block scopes in a subfile that you didn't open.