[WIP] HUGE error and file parsing refactor
- Refined InterpError to only cover interpreter errors
- Introduced TurnipTextContextlessError,
TurnipTextError to cover all errors
a) as emitted from the lexer/parser without file contents and
b) with all context needed to print error messages
- Moved file tracking logic out of interpreter and into TurnipTextParser
which takes contextless errors from lexer and interpreter and adds
context as needed.
- Refactored the interpreter to use structured token parsing.
Give it a mutable iterator of token and it will iterate through until
a InsertedFile is emitted,
at which point it bails out to the TurnipTextParser.
TurnipTextParser then calls it again with new data.
- Add file_idx to ParseSpan, so errors can span multiple files if needed
- Fixup all tests to work with new system