Make node traversal read-only in Python
After some thought, I realized it was both a bad idea and unnecessary to introduce arbitrary mutation on user-defined nodes.
There is no singleton protection, nodes might have structure that mutation interferes with
(e.g. for mutation to work easily you need to expose the contents of a node in an order, and allow arbitrary mutation of the list.
It could be that the number of nodes has semantic meaning or the specific indices have semantic meaning.)

This means there is no longer a need to have all nodes use the same field name for their iterables over contents, and we can use a base class with an exposer function instead.