Rename stateful -> in_doc, stateless -> pure_fmt, DocState -> DocEnv, FormatContext -> FmtEnv
The point of stateful vs stateless isn’t really state because Python is mutable at any time. @‌stateful funcs affect Document “state” or structure - they can add countable things and structure headers, and can only run when the document isn’t frozen i.e. as if they were invoked from inside the document - and @stateless funcs are intended only for pure formatting.

DocState.__dict__ is used as the execution environment for parsed Python code.

FormatContext was intended to be symmetrical with whatever DocState ended up as, and FmtEnv is symmetrical and the same length.

However, I decided to rename `doc` params to `doc_env` everywhere to disambiguate with `doc` as in `document`. This has the side effect of making it unsymmetrical. Might reverse that later.