Try to add an as_plain_text() method to Inlines
This was a mild failure because it doesn't integrate well with dfs. e.g. should an InlineScope implemenet as_plain_text by iterating through its children? but then if I do a DFS for Inline I will get as_plain_text for an InlineScope and it's children together. Need to think about fixing this.