StdAnchorPlugin -> AnchorEnv
Anchors are used by a lot of plugins, and served as the only part of those plugins that required poking the DocEnv without typechecking. This was error prone and broke things. Everything already assumed the StdAnchorPlugin was present/added by default, so making it a field in doc_env loses nothing and gains reliability.

This is backwards-incompatible because now any plugins that used doc_env.register_new_anchor etc. need to use doc_env.anchors.register...