Separate TeX content preambles from code preambles
Metadata like titles, authors, abstracts, etc. are usually defined in the document preamble.
Custom commands for e.g. formatting are also commonly defined in the document preamble.
Before, there was no way  to ensure that metadata was written after the formatting commands were defined. This technically works in practice with some things but I'm not super confident in it, so now you can specify what group you use when you add a preamble section and it will be ordered correctly.
As a bonus, this ensures the metadata is put right next to the \begin, which is more readable.