Replace EmitAs with CoerceBuilder, a typeclass which uses a function instead of a property
This is in line with the other builder types, makes it more idiomatic to define a builder-only class instead of a class deriving from both builder and (block|inline|header).
CoerceBuilder is built when it is passed to any coerce function, when it is emitted directly into the document, and can be returned by other builders before immediately being built.
Recursive CoerceBuilders are supported, with a fixed depth limit  currently defined to be 64 in CoerceBuilder::resolve_coercion.